+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `@todovue/tv-button` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
+## [1.1.3] - 2025-11-15
+### 🛠️ Changed
+- Removed CSS injection via `vite-plugin-css-injected-by-js` for the library build.
+- Styles are now served from a separate CSS file generated by Vite (`dist/*.css`) and must be imported explicitly in the consuming app (SPA or Nuxt).
+- Improved compatibility with SSR/SSG environments (especially Nuxt) by avoiding runtime CSS injection.
+
+---
 ## [1.1.2] - 2025-10-17
 ### 🐛 Fixed
 - Fixed CSS export to avoid the need for manual import wherever the component is used.
@@ -26,6 +33,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Updated Vite to `^7.0.0` to ensure compatibility with Node.js 20.19+.
 - Updated @vitejs/plugin-vue to `^6.0.0`.
 
+---
 ## [1.1.0] - 2025-05-08
 ### ✨ Added
 - Add icon-only button support with styling and documentation updates
@@ -35,8 +43,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Fixed button layout and icon positioning
 - Fixed tv-btn-text styles for improved accessibility and hover effects
 
+---
 ## [1.0.0] - 2025-05-05
-
 ### ✨ Added
 - Introduced `TvButton` as a reusable and fully customizable button component.
 - Support for multiple button variants: `default`, `outlined`, `text`, `icon-only`.
@@ -57,8 +65,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Ready-to-use as a standalone or global Vue component.
 - Scoped styles using SCSS.
 
-
 ---
+[1.1.3]: https://github.com/TODOvue/tv-button/pull/17/files
 [1.1.2]: https://github.com/TODOvue/tv-button/pull/16/files
 [1.1.1]: https://github.com/TODOvue/tv-button/pull/15/files
 [1.1.0]: https://github.com/TODOvue/tv-button/pull/13/files
