+++ README.md
@@ -5,8 +5,15 @@
 A flexible, framework‑agnostic Vue 3 button component with variants, sizes, icons, loading state, and customization utilities. Ship it in Single Page Apps or Server-Side Rendered (SSR) environments (e.g. Nuxt 3) with zero DOM assumptions.
 
 [![npm](https://img.shields.io/npm/v/@todovue/tv-button.svg)](https://www.npmjs.com/package/@todovue/tv-button)
+[![Netlify Status](https://api.netlify.com/api/v1/badges/3c413109-63aa-41d7-8126-a527435f5512/deploy-status)](https://app.netlify.com/projects/tv-button/deploys)
 [![npm downloads](https://img.shields.io/npm/dm/@todovue/tv-button.svg)](https://www.npmjs.com/package/@todovue/tv-button)
+[![npm total downloads](https://img.shields.io/npm/dt/@todovue/tv-button.svg)](https://www.npmjs.com/package/@todovue/tv-button)
 ![License](https://img.shields.io/github/license/TODOvue/tv-button)
+![Release Date](https://img.shields.io/github/release-date/TODOvue/tv-button)
+![Bundle Size](https://img.shields.io/bundlephobia/minzip/@todovue/tv-button)
+![Node Version](https://img.shields.io/node/v/@todovue/tv-button)
+![Last Commit](https://img.shields.io/github/last-commit/TODOvue/tv-button)
+![Stars](https://img.shields.io/github/stars/TODOvue/tv-button?style=social)
 
 > Demo: https://tv-button.netlify.app/
 
@@ -15,6 +22,7 @@ A flexible, framework‑agnostic Vue 3 button component with variants, sizes, ic
 - [Features](#features)
 - [Installation](#installation)
 - [Quick Start (SPA)](#quick-start-spa)
+- [Style usage](#style-usage)
 - [Nuxt 3 / SSR Usage](#nuxt-3--ssr-usage)
 - [Component Registration Options](#component-registration-options)
 - [Props](#props)
@@ -62,6 +70,7 @@ Global registration (main.js / main.ts):
 ```js
 import { createApp } from 'vue'
 import App from './App.vue'
+import '@todovue/tv-button/style.css'
 import TvButton from '@todovue/tv-button'
 
 createApp(App)
@@ -71,6 +80,7 @@ createApp(App)
 Local import inside a component:
 ```vue
 <script setup>
+import '@todovue/tv-button/style.css'
 import { TvButton } from '@todovue/tv-button'
 
 function onSubmit() {
@@ -81,6 +91,31 @@ function onSubmit() {
 <template>
   <TvButton success icon="check" @click-button="onSubmit">Submit</TvButton>
 </template>
+```
+
+---
+## Style usage
+
+### Vue 3 SPA with Vite
+```ts
+// main.ts
+import { createApp } from 'vue'
+import App from './App.vue'
+
+import '@todovue/tv-button/style.css'
+import { TvButton } from '@todovue/tv-button'
+
+const app = createApp(App)
+app.component('TvButton', TvButton)
+app.mount('#app')
+```
+
+### Nuxt 3/4
+```ts
+// nuxt.config.ts
+export default defineNuxtConfig({
+  css: ['@todovue/tv-button/style.css'],
+})
 ```
 
 ---
@@ -203,7 +238,7 @@ Loading state:
 ---
 ## SSR Notes
 - No direct DOM (`window` / `document`) access in source → safe for SSR.
-- Styles are now applied automatically when you import the library (no manual CSS import required).
+- Styles are now served from a separate CSS file generated by Vite (`dist/tv-button.css`). You need to import it explicitly in your app (SPA or Nuxt) using `@todovue/tv-button/style.css`.
 - SVG icons are bundled via Vite's `import.meta.glob` (works in Vite + Nuxt).
 
 ---
