+++ src/components/TvButton.vue
@@ -108,4 +108,4 @@ const buttonStyles = computed(() => ({
   </button>
 </template>
 
-<style scoped lang="scss" src="../assets/scss/style.scss"></style>
+<style></style>
