+++ src/demo/utils/demos/iconOnly.vue
