+++ src/demo/utils/demos/info.vue
