+++ src/demo/utils/demos/typeIcon.vue
