+++ src/demo/utils/demos/typeIconOutlinedCustom.vue
