+++ src/demo/utils/demos/typeIconOutlinedRemove.vue
