+++ src/style.scss
@@ -1,0 +1,2 @@
+@use "./assets/scss/style";
+
